+++ MyPluginInfo.cs
@@ -4,6 +4,6 @@ namespace PeakTextChat
     {
         public const string PLUGIN_GUID = "V8.TwitchChat";
         public const string PLUGIN_NAME = "TwitchChat";
-        public const string PLUGIN_VERSION = "1.0.3";
+        public const string PLUGIN_VERSION = "1.2.0";
     }
 }